requirements.txt
pyyaml>=6.0
requests>=2.25.0
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6
